Grammar, lore, and tone adjustment
Addresses a variety of grammatical errors in addition to beginning to
gradually shift tone and structure of dialogue from the dated
native-esque format to a more modern and appropriate form for the genre.
Removed and subsequently replaced all references to "purse", convoy",
and "'tis" and replaces them to fit more with lore, as well as a fact
made by the Tavern Traveler.